requirements.txt
Django==4.2
PyJWT==2.7.0
PyMySQL==1.0.3
tqdm==4.66.2
shortuuid==1.0.11
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2